+++ R/BF.print.R
@@ -46,6 +46,25 @@ print.BF <- function(x,
         cat("\n")
       }
     }
+    if(class(x$model)[1]=="rma.uni"){
+      cat("Posterior support for random effects","\n", sep = "")
+      cat("\n")
+      tau2out <- matrix(x$estimates[2,5])
+      row.names(tau2out) <- "tau2 (marema)"
+      colnames(tau2out) <- "Pr(>0)"
+      print(round(tau2out,digits))
+      cat("\n")
+    }
+    if(class(x$model)[1]=="cor_test"){
+      cat("Bartlett's test of sphericity (all correlations are zero):","\n", sep = "")
+      cat("\n")
+      if(nrow(x$PHP_Bartlett)>1){
+        print(round(x$PHP_Bartlett,digits))
+      }else{
+        print(round(x$PHP_Bartlett[1,],digits))
+      }
+      cat("\n")
+    }
 
   }else{
 
@@ -68,9 +87,9 @@ print.BF <- function(x,
 
     cat("\n")
     if(x$log==FALSE){
-      cat("Evidence matrix (BFs):")
+      cat("Bayes factors (rows versus columns):")
     }else{
-      cat("Evidence matrix (log BFs):")
+      cat("Log Bayes factors (rows versus columns):")
     }
     cat("\n")