+++ tests/testthat/test_metafor.R
@@ -10,9 +10,9 @@ test_that("metafor exploratory test", {
   set.seed(123)
   BF1 <- BF(res,BF.type="correlation",iter=1e3,cov.prob=.9)
   expect_equivalent(
-    round(BF1$BFtu_exploratory[,3],3),c(1.918,1.954), tolerance = .05
+    round(BF1$BFtu_exploratory[,3],3),c(1.9,1.9), tolerance = .05
   )
   expect_equivalent(
-    round(BF1$estimates[1:2,3],2),c(2.66,1.04), tolerance = .5
+    round(BF1$estimates[1:2,3],2),c(-.27,.34), tolerance = .5
   )
 })