+++ NEWS.md
@@ -1,6 +1,11 @@
+# BFpack 1.5.2
+
+* Bartlett's test for spericity (all correlations are zero) added.
+
 # BFpack 1.5.1
 
-* Version update as part of a publication on meta-analyses.
+* Version update of BF.rma.uni for a new publication on Bayes factors for meta-analyses.
+* Minor changes, bug fixes
 
 # BFpack 1.5.0
 
@@ -10,7 +15,7 @@
 # BFpack 1.4.1
 
 * Date: 2025-02-09
-* Bug fit of MCMC sampler for correlation test.
+* Bug fix of MCMC sampler for correlation test.
 
 # BFpack 1.4.0